requirements_python_full.txt
reflex
google-genai
python-dotenv
pydantic
psutil
pyautogui
opencv-python
